video-analyzer/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
python-multipart>=0.0.12
cryptography>=43.0.0
google-api-python-client>=2.150.0
google-auth>=2.35.0
openai>=1.50.0
anthropic>=0.37.0
google-generativeai>=0.8.0
mcp[server]>=1.0.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
aiosqlite>=0.20.0
python-jose[cryptography]>=3.3.0
moviepy>=2.0.0
Pillow>=10.4.0
